transactor programs: move test clients to io.mediachain.transactor.test
